+++ src/main/kotlin/io/openfuture/chain/crypto/domain/ECKey.kt
@@ -1,8 +1,11 @@
 package io.openfuture.chain.crypto.domain
 
+import io.openfuture.chain.util.HashUtils
 import org.bouncycastle.asn1.sec.SECNamedCurves
 import org.bouncycastle.asn1.sec.SECObjectIdentifiers
+import org.bouncycastle.pqc.math.linearalgebra.ByteUtils
 import java.math.BigInteger
+import java.util.*
 
 class ECKey(
     val private: BigInteger?,
@@ -15,17 +18,17 @@ class ECKey(
         private val curve = SECNamedCurves.getByOID(SECObjectIdentifiers.secp256k1)
     }
 
-    constructor(bytes: ByteArray): this(
+    constructor(bytes: ByteArray) : this(
         BigInteger(1, bytes),
         curve.g.multiply(BigInteger(1, bytes)).getEncoded(true)
     )
 
-    constructor(bytes: ByteArray, parent: ECKey): this(
+    constructor(bytes: ByteArray, parent: ECKey) : this(
         BigInteger(1, bytes).add(parent.private).mod(curve.n),
         curve.g.multiply(BigInteger(1, bytes).add(parent.private).mod(curve.n)).getEncoded(true)
     )
 
-    constructor(bytes: ByteArray, isPrivate: Boolean): this(
+    constructor(bytes: ByteArray, isPrivate: Boolean) : this(
         if (isPrivate) BigInteger(1, bytes) else null,
         if (isPrivate) curve.g.multiply(BigInteger(1, bytes)).getEncoded(true) else bytes
     )
@@ -44,6 +47,14 @@ class ECKey(
                 Math.max(0, PRIVATE_KEY_SIZE - privateInBytes.size), Math.min(PRIVATE_KEY_SIZE, privateInBytes.size))
             tmp
         }
+    }
+
+    fun isPrivateEmpty() = null == private
+
+    fun getAddress(): String {
+        val hash = HashUtils.keccakKeyHash(public)
+        val addressBytes = Arrays.copyOfRange(hash, 0, 20)
+        return "0x${ByteUtils.toHexString(addressBytes)}"
     }
 
 }